% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire_clean.r
\name{GLAD_clean}
\alias{GLAD_clean}
\title{Cleans One Specified Questionnaire}
\usage{
GLAD_clean(questionnaire, dat_list)
}
\arguments{
\item{questionnaire}{A questionnaire that is to be cleaned.}

\item{dat_list}{A named list of dataframes produced by 'GLAD_rawall'.}
}
\description{
Cleans the specified questionnaire and creates export.
}
