% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive.r
\name{GLAD_derive}
\alias{GLAD_derive}
\title{Generate Derived Variables For a Questionnaire Data Set.}
\usage{
GLAD_derive(data, googlesheet)
}
\arguments{
\item{data}{A dataframe produced by 'GLAD_clean' containing variables of
a questionnaire.}

\item{googlesheet}{A googlesheet produced by "GLAD_sheet" that contain
the dictionary sheet of the variables in 'data'.}
}
\value{
A dataframe with derived variable attached.
}
\description{
Generates derived variables with names and formulae specified in the
GLAD dictionary
}
