% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxi_fun.r
\name{GLAD_getdescr}
\alias{GLAD_getdescr}
\title{Get Descripton (Title) for Selected Variables}
\usage{
GLAD_getdescr(which, googlesheet)
}
\arguments{
\item{which}{An character vector indicating the items.}

\item{googlesheet}{A dataframe created by 'GLAD_sheet' that contains
corresponding dictionary information for 'data'.}
}
\value{
A named vector with names being the `Easy.name` or
`New.variable` and values being their descriptions.
}
\description{
Get Descripton (Title) for Selected Variables by specifying their names.
}
