% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.r
\name{GLAD_missing}
\alias{GLAD_missing}
\title{Examine Missingness of a Questionnaire}
\usage{
GLAD_missing(data)
}
\arguments{
\item{data}{A dataframe produced by 'GLAD_clean' containing variables of
a questionnaire.}
}
\value{
A named list of summary statistics and plots.
}
\description{
Produces various summary statistics and plots for missingness
examination of a questionnaire.
}
