% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glad_plot.r
\name{GLAD_plot}
\alias{GLAD_plot}
\title{Plot Variables in GLAD Data}
\usage{
GLAD_plot(
  data,
  var,
  title = "",
  include_outlier = TRUE,
  binwidth = "",
  googlesheet
)
}
\arguments{
\item{data}{The dataframe containing the variable to be plotted.}

\item{var}{A character string of one element specifying the variable to
be plotted}

\item{title}{A character string. Only applies to categorical
variables that allow multiple options to be selected.}

\item{include_outlier}{A logical. Whether to include cases exceeding the
minimum and maximum in the dictionary, if they exist in 'data'.}

\item{binwidth}{A number or a character string "FD". Only applies to
continuous variables. If specified as "FD", the Freedman-Diaconis rule
for deciding optimal binwidth will be applied.}

\item{googlesheet}{A googlesheet produced by "GLAD_sheet" that contain
the dictionary sheet of the variables in 'data'.}
}
\value{
For non-continuous variables, a single ggplot object. For continuous
variables, a named list of ggplot objects.
}
\description{
Generates plots for the specified variable in the GLAD data.
}
