% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_test.r
\name{GLAD_qplot}
\alias{GLAD_qplot}
\title{Generates Quantile Plot for a Continuous Variable}
\usage{
GLAD_qplot(data, var, googlesheet)
}
\arguments{
\item{data}{The dataframe containing the variable to be plotted.}

\item{var}{A character string of one element specifying the variable to
be plotted}

\item{googlesheet}{A dataframe produced by "GLAD_sheet" that contain
the dictionary sheet of the variables in 'data'.}
}
\value{
A ggplot object.
}
\description{
Generates quantile plots for a specified variable in the GLAD data.
}
