% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.r
\name{GLAD_read}
\alias{GLAD_read}
\title{Read In Raw Data}
\usage{
GLAD_read(path)
}
\arguments{
\item{path}{A character string specifying a folder containing the raw
files.}
}
\value{
A named list of dataframes each corresponding to a raw data file with
participant personal information removed.
}
\description{
Read in all the Qualtrics exported raw csv files in the specified path.
}
