% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googlesheet.r
\name{GLAD_sheet}
\alias{GLAD_sheet}
\title{Read in The Dictionary Sheet}
\usage{
GLAD_sheet(questionnaire)
}
\arguments{
\item{questionnaire}{The acronym of the questionnaire as one of the
sheet names in the GLAD dictionary googleshet.}
}
\value{
A dataframe containing information of the dictionary sheet.
}
\description{
Read in the googlesheet dictionary sheet for a specified questionnaire
}
