% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_test.r
\name{GLAD_vartest}
\alias{GLAD_vartest}
\title{Runs Variance Tests Between Genders}
\usage{
GLAD_vartest(data, var)
}
\arguments{
\item{data}{The dataframe containing the variable to be plotted.}

\item{var}{A character string of one element specifying the variable to
be tested.}
}
\description{
Runs several variance tests (Levene's test, Fligner's test and Barlett's
test) for comparing variance across gender.
}
