% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire_clean.r
\name{GLAD_cleanall}
\alias{GLAD_cleanall}
\title{Cleans All Questionnaires In a List}
\usage{
GLAD_cleanall(dat_list)
}
\arguments{
\item{dat_list}{A named list of dataframes produced by 'GLAD_rawall'.}
}
\description{
Cleans all the questionnaires in a list produced by 'GLAD_rawall'
and creates exports.
}
