% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_test.r
\name{GLAD_qplot}
\alias{GLAD_qplot}
\title{Generates Quantile Plot for a Continuous Variable}
\usage{
GLAD_qplot(data, var)
}
\arguments{
\item{data}{The dataframe containing the variable to be plotted.}

\item{var}{A character string of one element specifying the variable to
be plotted}
}
\value{
A quantile plot produced with 'ggplot2'
}
\description{
Generates quantile plots for a specified variable in the GLAD data.
}
