% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxi_fun.r
\name{GLAD_select}
\alias{GLAD_select}
\title{Exports Selected Variables From a Data Set}
\usage{
GLAD_select(data, which, googlesheet)
}
\arguments{
\item{data}{The dataframe containing the variables to be exported,
outputted by 'GLAD_clean' or 'GLAD_cleanall'.}

\item{which}{An character vector indicating the items to be export.}

\item{googlesheet}{A dataframe created by 'GLAD_sheet' that contains
corresponding dictionary information for 'data'.}
}
\description{
Exports selected variables from a data set by specifying their names.
}
