% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxi_fun.r
\name{GLAD_select}
\alias{GLAD_select}
\title{Exports Selected Variables From a Data Set}
\usage{
GLAD_select(clean_path, which)
}
\arguments{
\item{clean_path}{Path to where the exported files are.
outputted by 'GLAD_clean' or 'GLAD_cleanall'.}

\item{which}{Path to the *.txt file containing required variables.}
}
\description{
Exports selected variables specified through a *.txt file
}
