% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.r
\name{GLAD_select}
\alias{GLAD_select}
\title{Read in Selected Variables From Data sets to a list}
\usage{
GLAD_select(clean_path, export_path, person, which, format)
}
\arguments{
\item{clean_path}{Path to the folder containing the  exported files.}

\item{export_path}{Path to the data request folder, which should contain
subfolders each represent a person who requests the data}

\item{person}{The name of the person who requests the data}

\item{which}{A character vector specifying paths to the *.txt files
containing required variables. Each text file should correspond to and
has the name of a questionnaire. The variables within each text file
should be on seperate lines.}

\item{format}{Format of exported data. It should be one of c('rds',
'sav', 'sas', 'feather', 'csv')}
}
\description{
Read in selected variables specified through *.txt files from data sets
to a list
}
